requirements.txt
alembic==1.6.5
Babel==2.9.1
blinker==1.4
certifi==2021.5.30
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.13.4
email-validator==1.1.3
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
langdetect==1.0.9
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3

psycopg2==2.9.1
gunicorn==20.1.0